+++ lib/src/whitespace.dart
@@ -4,6 +4,30 @@
 
 library dart_style.src.whitespace;
 
+/// Constants for the number of spaces for various kinds of indentation.
+class Indent {
+  /// The number of spaces in a block or collection body.
+  static const block = 2;
+
+  /// The number of spaces in a single level of expression nesting.
+  static const expression = 4;
+
+  /// The ":" on a wrapped constructor initialization list.
+  static const constructorInitializer = 4;
+
+  /// The indentation for subsequent variables when a for loop defines multiple
+  /// variables that wrap, like:
+  ///
+  ///     for (var a = initializer,
+  ///             b = another,
+  ///             c = third;
+  ///         a + b + c < 100;
+  ///         a++) {
+  ///       ...
+  ///     }
+  static const loopVariable = 8;
+}
+
 /// The kind of pending whitespace that has been "written", but not actually
 /// physically output yet.
 ///